refactor app as per PR comments
