+++ src/app/app.module.ts
@@ -1,29 +1,17 @@
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
 import { NgModule } from '@angular/core';
 
 import { CoreModule } from './core/core.module';
 import { MaterialModule } from './material/material.module';
+import { SearchPageModule } from './search-page/search-page.module';
 
 import { AppComponent } from './app.component';
-import { InputComponent } from './input/input.component';
-import { ResultsComponent } from './results/results.component';
-import { MovieCardComponent } from './movie-card/movie-card.component';
-import { SearchPageComponent } from './search-page/search-page.component';
 
 @NgModule({
-  declarations: [AppComponent, InputComponent, ResultsComponent, MovieCardComponent, SearchPageComponent],
-  imports: [
-    BrowserAnimationsModule,
-    BrowserModule,
-    CoreModule,
-    FormsModule,
-    HttpClientModule,
-    MaterialModule,
-    ReactiveFormsModule
-  ],
+  declarations: [AppComponent],
+  imports: [BrowserAnimationsModule, BrowserModule, CoreModule, MaterialModule, SearchPageModule],
   providers: [],
   bootstrap: [AppComponent]
 })
