+++ src/app/search-page/input/input.component.scss
