+++ src/app/search-page/movie-card/movie-card.component.scss
