+++ src/app/search-page/results/results.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Movie } from '../core/models';
+import { Movie } from '../../core/models';
 
 @Component({
   selector: 'app-results',
@@ -9,6 +9,8 @@ import { Movie } from '../core/models';
 export class ResultsComponent {
   constructor() {}
 
-  /** This is used to receive value from parent component and render in template */
+  /**
+   * This is used to receive value from parent component and render in template
+   */
   @Input() public searchResult$!: Array<Movie>;
 }
